Reduced height of .speech-bubble-left and .speech-bubble-right ::after element to below 100% of container to avoid border overlap that looks like a gap in the speech bubble
Very, very excited to solve this issue today!!! The text box has an ::after element that conceals the inner edges of the triangular shape that makes up the speech bubble tongue. However, in intermediate cases, there will be a white overlay crossing over the speech bubble border. It looks like a gap in the border. When one pulls up the inspector, it promptly goes away (now known because the speech bubble is being resized to the point that the ::after content no longer overlaps). It was just a pixel, but it broke the continuity of the text box border and was difficult to identify just what was going on. Also, its inclusion in the developer's codebase presumed that it's configuration was required for the speech bubble configuration. But not so! The ::after element needs to be less than the height of the container. Setting this to 98% enables the speech bubble tongue to be low on the left or right of the speech bubble and still have its inner shape obscured without the chance of overlay. It will be important to remember as we begin customizing the speech bubbles in other roles such as flags and tooltips.