Add "tab-popup-pixel" to replace default background with nearly transparent image
Without an image selected, the link for the popup button in Smart Sparrow reveals a broken image-link icon and the alternate text. To turn the hotspot into a button with CSS-added text, I need to remove the image and replace it with something that doesn't look like it is there.