Update starburst-label with width: auto !important
The width property is set by the platform's authoring editor. When custom padding is applied to an element, the theme's CSS overrides any changes to width. But, it turns out that using "width: auto !important" resets the styling and allows for style.css specific styling to be applied to that element, such as border and border-radius. Happy Happy Happy!!!

This may also work as a solution for buttons and being able to apply previously cut-off box-shadows.