Add style update to video player popup
Add component titled twoThirdsVideoPlayerPopUp for playing video in a sequence. The play menu links are on the left and the video player occupies about two-thirds of the remaining area. Built through modifications to a popup component, customizations were required, including removing the 'close' button from the popup since display is controlled through the trapstate scripts.