Update videoPlayerPopUp close button with display: none
Update videoPlayerPopUp close button with display: none, because the audio auto-play issue when shifting window focus (eg. minimizing and coming back) is an issue in the LMS and is not affected by closing an open popup.

Options are either to leave as is (its rare) or turn off auto-play and require the user to click the button to show the popup containing the video, and then start / stop the video itself.